Handler and callback system
Handlers are expected to take data and a callback.

In the "req.on('end'" part, a handler is chosen based on the path using the
object router to map paths to handlers.  The handler is called and passed the
data of the request and a callback to call to terminate the request and send the
response.